Gradient Descent Basics and Matplolib
I have now figured out how gradient descent works. Gradient means derivative of the function and I used the basic equation of x = x - (derivative)*(learningRate). I looped through this a bunch of times and found the local minimum at an x and its value. I also Used matplot lib to show the function that I was using to get the gradient descent. I now see that gradient descent is useful in minimizing cost function as it will always settle into a local minimum however I do realize that a cost function such as MSE and multiple will result in a function with multiple local minimums and maximum making the function fall into a local minimum that is the best in accordance with its graph.